+++ inventree_quote_generator/__init__.py
@@ -1,3 +1,3 @@
 """InvenTree Quote Generator plugin package."""
 
-PLUGIN_VERSION = "0.2.1"
+PLUGIN_VERSION = "0.3.0"